requirements.txt
pyyaml
sentry-sdk
requests